requirements.txt
# 项目所使用的包
torch~=1.11.0+cu115
numpy~=1.21.6
torchvision~=0.12.0+cu115
opencv-python~=4.2.0.32
pillow~=9.4.0
scipy~=1.7.3
scikit-learn~=0.22
pandas~=1.3.5
tqdm~=4.65.0
matplotlib~=3.5.3
scikit-image~=0.19.3
six~=1.16.0
munch~=4.0.0
pydicom~=2.4.4
imageio~=2.19.3
vtk~=9.3.0